requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
Pillow>=8.2.0
tqdm>=4.62.0
matplotlib>=3.4.3
scipy>=1.7.1
tensorboard>=2.6.0
opencv-python>=4.5.3
einops>=0.3.2
transformers>=4.11.0  # 文本条件模型需要
scikit-learn>=0.24.2
pyyaml>=5.4.1
pandas>=1.3.3
wandb>=0.12.0  # 可选，用于实验跟踪
pytest>=6.2.5  # 用于测试 